pyproject.toml
[tool.poetry]
name = "nature-remo-streamlit"
version = "0.1.0"
description = ""
authors = ["miraium <miraium.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slack_sdk = "^3.21.3"
streamlit = "^1.36.0"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
requests = "^2.32.3"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
